fix: handle duplicates before adding to responses
